refactor(reposicao): extrai types/priority/constants de AdminRoutePlanner
Primeiro corte do god-component AdminRoutePlanner (1654 LoC, o maior).
Passo seguro (Codex order #1): extrai as folhas puras pra
src/components/reposicao/routePlanner/, zero mudança de lógica:

- types.ts     — StopType, PlanningMode, FilterPeriod, ManualFilter,
                 ManualCustomer, VisitStatus, RouteStop
- priority.ts  — enrichWithPriority (+ computeStopPriority interno)
- constants.ts — STOP_DURATION_MIN, PRIORITY_CONFIG, STOP_CONFIG

Página: 1654 → 1525 LoC. Remove imports lucide órfãos (ArrowUp/Right/
Down, que só o PRIORITY_CONFIG usava). Os 3 módulos entram no strict.

Os ~13 eslint no-explicit-any que aparecem são pré-existentes no corpo
da página (não introduzidos aqui; lint é informational no CI).

Verificado: tsc 0, typecheck:strict 0, eslint-unused 0, build OK, 392
testes passando.
